Add default value for collection options (#21)
This commit fixes an issue where no options are passed and a collection
exists. This passed undefined into a update function which then looked
for options to update, resulting in the following error:

```
TypeError: Cannot read property 'waitForSync' of undefined
```
This commit reproduces the error in a test and provides a default.